alert

+++ app/Http/Controllers/MahasiswaController.php
@@ -178,7 +178,9 @@ class MahasiswaController extends Controller
                 DB::rollBack();
                 return ResponseHelper::send('Tim tidak ditemukan.', null, 404);
             }
-
+            if ($team->projects->count() > 0) {
+                return ResponseHelper::send('Tim masih memiliki project. Silahkan hapus project terlebih dahulu', null, 400);
+            }
             $team->delete(); // Ini saja sudah cukup! Database akan menghapus anggota_teams.
 
             DB::commit();
@@ -296,7 +298,7 @@ class MahasiswaController extends Controller
             }
 
             if ($member->projects->count() > 0) {
-                return ResponseHelper::send('Mahasiswa masih punya project', null, 400);
+                return ResponseHelper::send('Mahasiswa masih punya project. Silahkan hapus project terlebih dahulu', null, 400);
             }
 
             // Hapus foto jika ada
